pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "wcosmo"
authors = [
    {name = "Colm Talbot", email = "talbotcolm@gmail.com"},
    {name = "Amanda Farah"},
]
description = "A package for cosmology calculations with arbitrary numpy-like APIs"
readme = "README.rst"
requires-python = ">=3.10"
license = {text = "MIT"}
classifiers = [
    "Framework :: Django",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "numpy",
    "scipy",
    "astropy>=6.1",
]
dynamic = ["version"]

[project.optional-dependencies]
test = [
    "pytest-cov",
    "gwpopulation",
    "jax>=0.4.16",
    "unxt",
]

[tool.setuptools]
packages = ["wcosmo"]

[tool.setuptools_scm]
write_to = "wcosmo/_version.py"

[project.entry-points."gwpopulation.xp"]
wcosmo = "wcosmo.wcosmo"
wcosmo-analytic = "wcosmo.analytic"
wcosmo-astropy = "wcosmo.astropy"
wcosmo-integrate = "wcosmo.integrate"
wcosmo-taylor = "wcosmo.taylor"
wcosmo-utils = "wcosmo.utils"

[project.entry-points."gwpopulation.scs"]
wcosmo-analytic = "wcosmo.analytic"

[project.entry-points."gwpopulation.other"]
wcosmo-taylor = "wcosmo.taylor:scipy.linalg.toeplitz"
